+++ src/main/kotlin/kr/co/archan/reflect/global/exception/base/ProductException.kt
@@ -1,3 +1,3 @@
-package kr.co.archan.reflect.global.exception
+package kr.co.archan.reflect.global.exception.base
 
 abstract class ProductException(message: String) : RuntimeException(message)
