+++ src/main/kotlin/kr/co/archan/reflect/global/util/extension/Preconditions.kt
@@ -1,6 +1,6 @@
 package kr.co.archan.reflect.global.util.extension
 
-import kr.co.archan.reflect.global.exception.ProductException
+import kr.co.archan.reflect.global.exception.base.ProductException
 
 inline fun requireAndThrowProductException(value: Boolean, ex: () -> ProductException) {
     if (!value) throw ex()
