+++ src/main/kotlin/kr/co/archan/reflect/member/exception/NoMemberException.kt
@@ -1,6 +1,6 @@
 package kr.co.archan.reflect.member.exception
 
-import kr.co.archan.reflect.global.exception.ProductException
+import kr.co.archan.reflect.global.exception.base.ProductException
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
